+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # TelegramBotAPI changelog
 
+## 2.1.1
+
+* `API`:
+  * Now it is possible to edit any message text via `editMessageText` (but with warning)
+  * Now it is possible to edit any message caption via `editMessageCaption` (but with warning)
+  * Media message caption edit method (`editMessageCaption`) now returns the message with the same generic type (as it must be in telegram system)
+  * New extensions `TelegramBot#edit` has been added for all possible editions types
+  * New extensions `TelegramBot#send` has been added for all possible sending types
+  * New extensions `TelegramBot#delete` has been added
+* `Versions`:
+    * `MicroUtils`: `0.11.3` -> `0.11.6`
+
 ## 2.1.0
 
 __This update contains including of [Telegram Bot API 6.1](https://core.telegram.org/bots/api-changelog#june-20-2022)__